don't download files that are already downloaded

+++ src/main.rs
@@ -196,21 +196,26 @@ impl SrcLoc {
 }
 
 fn download_crate(krate: &Crat) -> PathBuf {
-    println!("Downloading {}-{} ...", krate.name, krate.version);
-    let mut url: String = String::from("https://crates.io/api/v1/crates/");
-    url.push_str(krate.name);
-    url.push_str("/");
-    url.push_str(&krate.version.to_string());
-    url.push_str("/");
-    url.push_str("download");
-
-    let mut req = reqwest::get(url.as_str())
-        .unwrap_or_else(|_| panic!("Failed to downloadCrate {:?}", krate));
     let filename = format!("{}-{}.crate", krate.name, krate.version.to_string());
     let dest_path = PathBuf::from("downloads/").join(filename);
-    let mut dest_file = std::fs::File::create(&dest_path).unwrap();
 
-    std::io::copy(&mut req, &mut dest_file).unwrap();
+    // don't download files we already have
+    if PathBuf::from(&dest_path).exists() {
+        println!("Downloading {}-{} ...", krate.name, krate.version);
+        let mut url: String = String::from("https://crates.io/api/v1/crates/");
+        url.push_str(krate.name);
+        url.push_str("/");
+        url.push_str(&krate.version.to_string());
+        url.push_str("/");
+        url.push_str("download");
+
+        let mut req = reqwest::get(url.as_str())
+            .unwrap_or_else(|_| panic!("Failed to downloadCrate {:?}", krate));
+
+        let mut dest_file = std::fs::File::create(&dest_path).unwrap();
+
+        std::io::copy(&mut req, &mut dest_file).unwrap();
+    }
     dest_path
 }
